update: withContent convert to tsx
